refactor: extract ThreadSafeAccess to common
