+++ basyx.common/pom.xml
@@ -20,6 +20,7 @@
 		<module>basyx.mongocore</module>
 		<module>basyx.authorization</module>
 		<module>basyx.client</module>
+		<module>basyx.backend</module>
 		<module>basyx.backend.inmemory.core</module>
 		<module>basyx.filerepository-backend</module>
 		<module>basyx.filerepository-backend-inmemory</module>
