+++ basyx.submodelregistry/basyx.submodelregistry-service-inmemory-storage/pom.xml
@@ -19,6 +19,10 @@
 	<dependencies>
 		<dependency>
 			<groupId>org.eclipse.digitaltwin.basyx</groupId>
+			<artifactId>basyx.backend</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.digitaltwin.basyx</groupId>
 			<artifactId>basyx.submodelregistry-service</artifactId>
 		</dependency>
 		<dependency>
