+++ basyx.submodelregistry/basyx.submodelregistry-service-inmemory-storage/src/main/java/org/eclipse/digitaltwin/basyx/submodelregistry/service/storage/memory/ThreadSafeSubmodelRegistryStorageDecorator.java
@@ -27,6 +27,7 @@ package org.eclipse.digitaltwin.basyx.submodelregistry.service.storage.memory;
 import java.util.List;
 import java.util.Set;
 
+import org.eclipse.digitaltwin.basyx.common.backend.ThreadSafeAccess;
 import org.eclipse.digitaltwin.basyx.core.pagination.CursorResult;
 import org.eclipse.digitaltwin.basyx.core.pagination.PaginationInfo;
 import org.eclipse.digitaltwin.basyx.submodelregistry.model.SubmodelDescriptor;
